[FIX] Solved a bug where final eval results evaluated overlapped tokens twice when using stride in sliding window.
This caused F1 scores to appear better or worse than the actual results.

If the scores are above 50% it's more probable that the overlapped entities are correctly classified and thus scores are boosted. In case scores are below 50%, your scores are going to be lower due to an incorrect classified entity being counted twice (or more depending on the overlapped percentage)